feat: support build fallback entry
