+++ packages/ice/src/middlewares/renderMiddleware.ts
@@ -43,6 +43,7 @@ export default function createRenderMiddleware(options: Options): Middleware {
       // Ignore static resources.
       !isStaticResources.test(req.path)) {
       const serverModule = await excuteServerEntry();
+      console.log('serverModule', serverModule);
       if (serverModule) {
         const requestContext: ServerContext = {
           req,
