+++ packages/ice/templates/core/entry.server.ts.ejs
@@ -1,5 +1,9 @@
 import './env.server';
-import * as runtime from '@ice/runtime/server';
+<% if (hydrate) {-%>
+import { getAppConfig, renderToHTML as renderAppToHTML, renderToResponse as renderAppToResponse } from '@ice/runtime/server';
+<% } else { -%>
+import { getAppConfig, getDocumentResponse as renderAppToHTML, renderDocumentToResponse as renderAppToResponse } from '@ice/runtime/server';
+<% }-%>
 <% if (hydrate) {-%>
 import { commons, statics } from './runtime-modules';
 <% }-%>
@@ -29,10 +33,11 @@ const createRoutes = () => routesManifest;
 const runtimeModules = { commons, statics };
 
 const getRouterBasename = () => {
-  const appConfig = runtime.getAppConfig(app);
+  const appConfig = getAppConfig(app);
   return appConfig?.router?.basename ?? <%- basename %> ?? '';
 }
 
+<% if (hydrate) {-%>
 const setRuntimeEnv = (renderMode) => {
   if (renderMode === 'SSG') {
     process.env.ICE_CORE_SSG = 'true';
@@ -40,6 +45,7 @@ const setRuntimeEnv = (renderMode) => {
     process.env.ICE_CORE_SSR = 'true';
   }
 }
+<% } -%>
 
 interface RenderOptions {
   documentOnly?: boolean;
@@ -54,19 +60,21 @@ interface RenderOptions {
 }
 
 export async function renderToHTML(requestContext, options: RenderOptions = {}) {
+<% if (hydrate) {-%>
   const { renderMode = 'SSR' } = options;
   setRuntimeEnv(renderMode);
-
+<% }-%>
   const mergedOptions = mergeOptions(options);
-  return await runtime.renderToHTML(requestContext, mergedOptions);
+  return await renderAppToHTML(requestContext, mergedOptions);
 }
 
 export async function renderToResponse(requestContext, options: RenderOptions = {}) {
+<% if (hydrate) {-%>
   const { renderMode = 'SSR' } = options;
   setRuntimeEnv(renderMode);
-
+<% }-%>
   const mergedOptions = mergeOptions(options);
-  return runtime.renderToResponse(requestContext, mergedOptions);
+  return renderAppToResponse(requestContext, mergedOptions);
 }
 
 function mergeOptions(options) {
