+++ packages/runtime/src/index.server.ts
@@ -1,2 +1,3 @@
 export { renderToResponse, renderToHTML } from './runServerApp.js';
+export { renderToResponse as renderDocumentToResponse, getDocumentResponse } from './renderDocument.js';
 export * from './index.js';
